pyproject.toml
# pyproject.toml
# Python project configuration for Fesch Agent

[project]
name = "fesch-agent"
version = "0.1.0"
description = "Polyglot AI Agent - Rust + Python + C++"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["brain"]
